Moved `chrome.js` to `utils/chrome.ts`
