+++ src/discogs/app/metadata.js
@@ -1,6 +1,6 @@
 import { Release } from "../../app/release.js";
 import { keywordsToDiscogsGenres, keywordsToDiscogsStyles } from "../../bandcamp/modules/bandcamp.js";
-import { getExtensionManifest } from "../../modules/chrome.js";
+import { getExtensionManifest } from "../../utils/chrome";
 import { generateSubmissionNotes } from "../modules/discogs.js";
 import { getDiscogsDateValue } from "./utils.js";
