+++ src/popup/helpers.js
@@ -1,7 +1,7 @@
 import { Release, ReleaseItem } from "../app/release.js";
 import { Metadata } from "../discogs/app/metadata.js";
 import { getSearchDiscogsReleaseUrl } from "../discogs/modules/discogs.js";
-import { chromeSendMessageToCurrentTab } from "../modules/chrome.js";
+import { chromeSendMessageToCurrentTab } from "../utils/chrome";
 import { createIconLink, disable, enable, getDataAttribute, hasDataAttribute, setDataAttribute } from "../modules/html.js";
 import { generateKeyForReleaseItem } from "../modules/key-generator.js";
 import { convertToAlias, isArray, isObject, isString } from "../modules/utils.js";
